incz-back/requirements.txt
djangorestframework
django
django-cors-headers
django-filter
django-rest-swagger